Fixes a few of the slide animations
Fixes the case where, while in the demo, you're viewing the right side panel and you drag the center panel almost to the end. The animation moves too fast (in my opinion).

Fixes a case where bounce is disabled and you get to a side panel by way of the showRightPanelAnimated:YES method. When you hide the center panel (so that you can go deeper into the navigation), the animation to hide the center panel is extremely fast. This could be a very edge case, as it doesn't happen in the demo.